+++ app/Models/SetInfo.php
@@ -4,7 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Relations\Pivot;
 
-class setInfo extends Pivot
+class SetInfo extends Pivot
 {
     //
 }